refactor: add missing ng prefix to the angular module name
replace module name from 'ovh-angular-actions-menu' to 'ngOvhActionsMenu'

BREAKING CHANGE: module is now named as `ngOvhActionsMenu`
