+++ src/template.html
@@ -1,7 +1,7 @@
 <button type="button"
         class="actions-menu-button"
         data-ng-class="{ 'menu-open' : $ctrl.popoverSettings.isOpen }"
-        data-responsive-popover="'ovh-angular-actions-menu-inner.html'"
+        data-responsive-popover="'template-inner.html'"
         data-popover-class="{{ $ctrl.popoverSettings.class ? $ctrl.popoverSettings.class + ' ovh-angular-actions-menu' : 'ovh-angular-actions-menu' }}"
         data-popover-placement="{{ $ctrl.popoverSettings.placement }}"
         data-popover-trigger="{{ $ctrl.popoverSettings.trigger }}"
